requirements.txt
appdirs==1.4.3
Babel==2.4.0
bcrypt==3.1.3
cffi==1.10.0
Django==1.10.6
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-extensions==1.7.7
django-phonenumber-field==1.3.0
packaging==16.8
phonenumberslite==8.5.1
pycparser==2.17
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
sqlparse==0.2.3
